@@ -8,7 +8,7 @@ type Templates struct {
 }
 
 type Config struct {
-	Data      map[string]map[interface{}]interface{} `yaml:"data,omitempty"`
-	Include   map[string]string                      `yaml:"include,omitempty"`
-	Templates Templates                              `yaml:"templates,omitempty"`
+	Data      map[string]map[any]any `yaml:"data,omitempty"`
+	Include   map[string]string      `yaml:"include,omitempty"`
+	Templates Templates              `yaml:"templates,omitempty"`
 }
