@@ -70,21 +70,7 @@ func (tm *Model) isExcluded(source string) bool {
 	return false
 }
 
-func (tm *Model) isIncluded(source string) bool {
-	for _, incl := range tm.Config.Templates.Includes {
-		regex := regexp.MustCompile(incl)
-		if regex.MatchString(source) {
-			return true
-		}
-	}
-	return false
-}
-
 func (tm *Model) Generate() error {
-	if len(tm.Config.Templates.Excludes) > 0 && len(tm.Config.Templates.Includes) > 0 {
-		println("Both excludes and includes are defined in config file, so only includes will be used.")
-	}
-
 	osErr := os.MkdirAll(tm.TargetPath, os.ModePerm)
 	if osErr != nil {
 		return osErr
@@ -121,52 +107,57 @@ func applyTemplates(
 	tm *Model,
 	relSource string,
 	info os.FileInfo,
-	context map[string]map[interface{}]interface{},
+	context map[string]map[any]any,
 	err error,
 ) error {
-	if !tm.isExcluded(relSource) && tm.isIncluded(relSource) {
-		rel, err := filepath.Rel(tm.SourcePath, relSource)
-		if err != nil {
+	if tm.isExcluded(relSource) {
+		return err
+	}
+
+	if info.IsDir() {
+		return err
+	}
+
+	rel, err := filepath.Rel(tm.SourcePath, relSource)
+	if err != nil {
+		return err
+	}
+
+	currentTarget := filepath.Join(tm.TargetPath, rel)
+
+	gen := NewGenerator(
+		relSource,
+		currentTarget,
+		context,
+	)
+
+	realTarget, fErr := gen.processFilename(tm)
+	if fErr != nil { //maybe we should fail back to real name instead?
+		return fErr
+	}
+
+	gen.updateTarget(realTarget)
+
+	currentTargetDir := filepath.Dir(realTarget)
+
+	if _, err := os.Stat(currentTargetDir); os.IsNotExist(err) {
+		if err := os.MkdirAll(currentTargetDir, os.ModePerm); err != nil {
 			return err
 		}
-		currentTarget := filepath.Join(tm.TargetPath, rel)
-		if !info.IsDir() {
-			tmplSuffix := tm.Suffix
+	}
 
-			gen := NewGenerator(
-				relSource,
-				currentTarget,
-				context,
-			)
-
-			realTarget, fErr := gen.processFilename(tm)
-			if fErr != nil { //maybe we should fail back to real name instead?
-				return fErr
-			}
-
-			gen.updateTarget(realTarget)
-
-			currentTargetDir := filepath.Dir(realTarget)
-
-			if _, err := os.Stat(currentTargetDir); os.IsNotExist(err) {
-				if err := os.MkdirAll(currentTargetDir, os.ModePerm); err != nil {
-					return err
-				}
-			}
-
-			if strings.HasSuffix(info.Name(), tmplSuffix) {
-				content, cErr := gen.processFile()
-				if cErr != nil {
-					return cErr
-				}
-
-				return io.CopyBufferToFile(content, relSource, realTarget)
-			} else {
-				if _, err := io.CopyFromSourceToTarget(relSource, realTarget); err != nil {
-					return err
-				}
-			}
+	if strings.HasSuffix(info.Name(), tm.Suffix) {
+		content, cErr := gen.processFile()
+		if cErr != nil {
+			return cErr
 		}
+
+		return io.CopyBufferToFile(content, relSource, realTarget)
+	}
+
+	_, err = io.CopyFromSourceToTarget(relSource, realTarget)
+	if err != nil {
+		return err
 	}
 
 	return err
